Generating all the components models automatically using new way of doing things :)
