+++ licket-framework/src/main/java/org/licket/core/view/container/AbstractLicketContainer.java
@@ -1,14 +1,27 @@
 package org.licket.core.view.container;
 
+import static com.fasterxml.jackson.core.JsonGenerator.Feature.QUOTE_FIELD_NAMES;
 import static com.google.common.collect.Lists.newArrayList;
+import static java.lang.String.format;
 import static org.licket.core.model.LicketModel.emptyModel;
+
+import java.io.IOException;
+import java.io.Reader;
+import java.io.StringReader;
 import java.util.List;
 import java.util.function.Predicate;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.licket.core.id.CompositeId;
 import org.licket.core.model.LicketModel;
 import org.licket.core.view.AbstractLicketComponent;
 import org.licket.core.view.ComponentView;
 import org.licket.core.view.LicketComponent;
+import org.licket.core.view.hippo.testing.annotation.Name;
+import org.licket.framework.hippo.ObjectLiteralBuilder;
+import org.mozilla.javascript.Parser;
+import org.mozilla.javascript.ast.AstRoot;
+import org.mozilla.javascript.ast.ObjectLiteral;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -20,6 +33,9 @@ public abstract class AbstractLicketContainer<T> extends AbstractLicketComponent
     private static final Logger LOGGER = LoggerFactory.getLogger(AbstractLicketContainer.class);
     private List<LicketComponentContainer<?>> branches = newArrayList();
     private List<LicketComponent<?>> leaves = newArrayList();
+
+    @Name("model")
+    private ObjectLiteralBuilder modelProperty = ObjectLiteralBuilder.objectLiteral();
 
     public AbstractLicketContainer(String id, Class<T> modelClass, ComponentView view) {
         this(id, modelClass, emptyModel(), view);
@@ -48,7 +64,35 @@ public abstract class AbstractLicketContainer<T> extends AbstractLicketComponent
     protected final void onInitialize() {
         branches.forEach(LicketComponent::initialize);
         leaves.forEach(LicketComponent::initialize);
+
         onInitializeContainer();
+        generateComponentModel();
+    }
+
+    private void generateComponentModel() {
+        try {
+            ObjectMapper mapper = new ObjectMapper();
+            mapper.configure(QUOTE_FIELD_NAMES, false);
+
+            // serialize component model to string json
+            String modelStringValue = mapper.writeValueAsString(getComponentModel().get());
+
+            // parse model declaration object literal
+            AstRoot astRoot = new Parser().parse(modelObjectLiteralReader(modelStringValue), "test.js", 0);
+            astRoot.visitAll(node -> {
+                if (node instanceof ObjectLiteral) {
+                    modelProperty.fromObjectLiteral((ObjectLiteral) node);
+                    return false;
+                }
+                return true;
+            });
+        } catch (IOException e) {
+            LOGGER.error("An error occurred while creating Angular class constructor.", e);
+        }
+    }
+
+    private Reader modelObjectLiteralReader(String modelStringValue) {
+        return new StringReader(format("model = %s", modelStringValue));
     }
 
     protected void onInitializeContainer() {}
