+++ licket-framework/src/main/java/org/licket/core/view/hippo/testing/ngclass/AngularClassPropertiesDecorator.java
@@ -1,8 +1,7 @@
 package org.licket.core.view.hippo.testing.ngclass;
 
 import org.licket.core.view.hippo.testing.annotation.Name;
-import org.licket.framework.hippo.ObjectLiteralBuilder;
-import org.licket.framework.hippo.ObjectPropertyBuilder;
+import org.licket.framework.hippo.*;
 import org.springframework.util.ReflectionUtils;
 
 import java.lang.reflect.Field;
@@ -12,8 +11,13 @@ import static com.google.common.base.Predicates.assignableFrom;
 import static com.google.common.base.Predicates.notNull;
 import static org.apache.commons.lang.StringUtils.trimToNull;
 import static org.joor.Reflect.on;
+import static org.licket.framework.hippo.AssignmentBuilder.assignment;
+import static org.licket.framework.hippo.ExpressionStatementBuilder.expressionStatement;
+import static org.licket.framework.hippo.KeywordLiteralBuilder.thisLiteral;
+import static org.licket.framework.hippo.NameBuilder.name;
 import static org.licket.framework.hippo.ObjectLiteralBuilder.objectLiteral;
 import static org.licket.framework.hippo.ObjectPropertyBuilder.propertyBuilder;
+import static org.licket.framework.hippo.PropertyNameBuilder.property;
 import static org.springframework.util.ReflectionUtils.doWithFields;
 
 /**
@@ -31,25 +35,26 @@ public class AngularClassPropertiesDecorator {
         this.angularClass = angularClass;
     }
 
-    public ObjectLiteralBuilder decorate(ObjectLiteralBuilder objectLiteral) {
+    public BlockBuilder decorate(BlockBuilder block) {
         doWithFields(angularClass.getClass(), field -> {
             doOnAccessible(field, accessibleField -> {
-                ObjectPropertyBuilder objectProperty = (ObjectPropertyBuilder) accessibleField.get(angularClass);
+                ObjectLiteralBuilder objectProperty = (ObjectLiteralBuilder) accessibleField.get(angularClass);
                 if (objectProperty == null) {
-                    objectProperty = propertyBuilder().value(objectLiteral());
+                    objectProperty = objectLiteral();
                 }
                 Name customName = accessibleField.getAnnotation(Name.class);
-                if (customName != null) {
-                    objectProperty.name(firstNonNull(trimToNull(customName.value()), accessibleField.getName()));
-                } else {
-                    objectProperty.name(accessibleField.getName());
-                }
-                objectLiteral.objectProperty(objectProperty);
+                PropertyNameBuilder assignmentProperty = property(thisLiteral(),
+                        name(firstNonNull(trimToNull(customName.value()), accessibleField.getName())));
+
+
+                block.appendStatement(expressionStatement(assignment()
+                        .left(assignmentProperty)
+                        .right(objectProperty)));
             });
         }, field -> {
-            return notNull().apply(on(field).get()) && assignableFrom(ObjectPropertyBuilder.class).apply(field.getType());
+            return notNull().apply(on(field).get()) && assignableFrom(ObjectLiteralBuilder.class).apply(field.getType());
         });
-        return objectLiteral;
+        return block;
     }
 
     private void doOnAccessible(Field field, ReflectionUtils.FieldCallback fieldConsumer)
